remove some code, and move some code
